# Reduce compression memory use with map[string]uint16