@miekg 2018-12-01
[ Quoting <notifications@github.com> in "[miekg/dns] Reduce compression memo..." ]
>This PR switches the packing code to use a `map[string]uint16` for storing compression pointers while keeping support for external callers who pass in a `map[string]int`. This reduces the per-entry memory consumption of the compression map by 25%. (`(16+2)/(16+8) = 0.75`).
>
>`compressionMap` is passed quite specifically as a value rather than a pointer because `RR.pack` is an interface method call which would cause it to escape and be heap allocated if it were a pointer. As maps are internally represented by a pointer to the runtime's `hmap` struct, the size of `compressionMap` is just 16-bytes anyway.
>
>The benchmarks show a notable improvement in both memory use (alloc/op) and performance (time/op) for packing messages where multiple records are present.
>
>```
>name                    old time/op    new time/op    delta
>MsgLengthPack-12          1.41µs ± 9%    1.46µs ±10%     ~     (p=0.353 n=10+10)
>PackDomainName-12          128ns ± 0%     137ns ± 1%   +6.72%  (p=0.000 n=6+10)
>PackA-12                  39.0ns ± 1%    40.1ns ± 1%   +2.69%  (p=0.000 n=8+8)
>PackMX-12                 69.6ns ± 1%    72.4ns ± 1%   +4.05%  (p=0.000 n=10+9)
>PackAAAAA-12              37.9ns ± 1%    40.0ns ± 1%   +5.68%  (p=0.000 n=10+9)
>PackMsg-12                1.01µs ± 5%    0.95µs ± 1%   -5.85%  (p=0.000 n=10+10)
>PackMsgOnlyQuestion-12     168ns ± 0%     176ns ± 1%   +4.94%  (p=0.000 n=6+10)
>PackMsgMassive-12         67.8µs ±10%    55.1µs ±10%  -18.70%  (p=0.000 n=10+10)
>
>name                    old alloc/op   new alloc/op   delta
>MsgLengthPack-12            576B ± 0%      528B ± 0%   -8.33%  (p=0.000 n=10+10)
>PackDomainName-12          0.00B          0.00B          ~     (all equal)
>PackA-12                   0.00B          0.00B          ~     (all equal)
>PackMX-12                  0.00B          0.00B          ~     (all equal)
>PackAAAAA-12               0.00B          0.00B          ~     (all equal)
>PackMsg-12                  256B ± 0%      208B ± 0%  -18.75%  (p=0.000 n=10+10)
>PackMsgOnlyQuestion-12     0.00B          0.00B          ~     (all equal)
>PackMsgMassive-12         23.1kB ± 0%    19.0kB ± 0%  -17.88%  (p=0.000 n=10+10)
>
>name                    old allocs/op  new allocs/op  delta
>MsgLengthPack-12            3.00 ± 0%      3.00 ± 0%     ~     (all equal)
>PackDomainName-12           0.00           0.00          ~     (all equal)
>PackA-12                    0.00           0.00          ~     (all equal)
>PackMX-12                   0.00           0.00          ~     (all equal)
>PackAAAAA-12                0.00           0.00          ~     (all equal)
>PackMsg-12                  2.00 ± 0%      2.00 ± 0%     ~     (all equal)
>PackMsgOnlyQuestion-12      0.00           0.00          ~     (all equal)
>PackMsgMassive-12           12.0 ± 0%      12.0 ± 0%     ~     (all equal)
>```

Looking at these benchmarks are the skewed to larger messages? 'Cause most DNS 
message are still small (less than 10 RRs)