@tmthrgd 2018-12-01
Only `PackMsgMassive` really exercises the compression code that well. The other benchmarks usually only include a single name with three labels multiple times, meaning the compression map has only three entries and it barely has any effect on the overall performance. I think real world results would be more positive than anything but `PackMsgMassive`.

If you look at the benchmarks, the increases are on the order of a few nanoseconds (and I believe come solely down to an extra few `nil` checks), while `PackMsgMassive` shows an improvement of over ten microseconds. I think something like Consul(?) would see a strong improvement here.